@@ -1,17 +1,12 @@
 import {AddressInfo} from 'net';
-import * as http from 'http';
-import * as path from 'path';
-import * as fs from 'fs';
-import {promisify} from 'util';
-import * as marked from 'marked';
-import * as mime from 'mime';
+import http from 'http';
+import path from 'path';
+import {promises as fs} from 'fs';
+import {marked} from 'marked';
+import mime from 'mime';
 import {URL} from 'url';
-import escape = require('escape-html');
-import enableDestroy = require('server-destroy');
-
-const readFile = promisify(fs.readFile);
-const stat = promisify(fs.stat);
-const readdir = promisify(fs.readdir);
+import escape from 'escape-html';
+import enableDestroy from 'server-destroy';
 
 export interface WebServerOptions {
   // The local path that should be mounted as a static web server
@@ -66,7 +61,7 @@ async function handleRequest(
     options.directoryListing && localPath.endsWith(`${path.sep}index.html`);
 
   try {
-    const stats = await stat(localPath);
+    const stats = await fs.stat(localPath);
     const isDirectory = stats.isDirectory();
     if (isDirectory) {
       // this means we got a path with no / at the end!
@@ -78,14 +73,15 @@ async function handleRequest(
       res.end(doc);
       return;
     }
-  } catch (err) {
+  } catch (e) {
+    const err = e as Error;
     if (!maybeListing) {
       return return404(res, err);
     }
   }
 
   try {
-    let data = await readFile(localPath, {encoding: 'utf8'});
+    let data = await fs.readFile(localPath, {encoding: 'utf8'});
     let mimeType = mime.getType(localPath);
     const isMarkdown = req.url?.toLocaleLowerCase().endsWith('.md');
     if (isMarkdown && options.markdown) {
@@ -96,10 +92,10 @@ async function handleRequest(
     res.setHeader('Content-Length', Buffer.byteLength(data));
     res.writeHead(200);
     res.end(data);
-  } catch (err) {
+  } catch (e) {
     if (maybeListing) {
       try {
-        const files = await readdir(originalPath);
+        const files = await fs.readdir(originalPath);
         const fileList = files
           .filter(f => escape(f))
           .map(f => `<li><a href="${f}">${f}</a></li>`)
@@ -108,10 +104,12 @@ async function handleRequest(
         res.writeHead(200);
         res.end(data);
         return;
-      } catch (err) {
+      } catch (e) {
+        const err = e as Error;
         return return404(res, err);
       }
     } else {
+      const err = e as Error;
       return return404(res, err);
     }
   }
